fix(Algorithm): modify the console output, will be further deprecated

+++ src/main/java/Algorithm/TSP1.java
@@ -8,7 +8,7 @@ public class TSP1 extends TemplateTSP {
 
 
 	@Override
-	protected int bound(long currentVertex, Collection<Long> unvisited) {
+	protected double bound(long currentVertex, Collection<Long> unvisited) {
 		return 0;
 	}
